feat:#222 유틸 함수에 책임 전가
